mqmdatatypes.cpp + mqmmixture.cpp: explicit conversions to avoid compiler warnings
Former-commit-id: bbd67bb9c15d115bf1b4c9e94ab3b38febe449f9

+++ src/mqmdatatypes.cpp
@@ -108,7 +108,7 @@ void change_coding(int *Nmark, int *Nind, int **Geno, MQMMarkerMatrix markers, c
 
 void *calloc_init(size_t num, size_t size) {
   void *buf;
-  buf = S_alloc(num,size);
+  buf = S_alloc(num,(int)size);
   return buf;
 }
